Changes in collection and fixing admin notifications
