+++ wazimap_ng/points/admin/__init__.py
@@ -2,3 +2,4 @@ from .theme_admin import ThemeAdmin
 from .location_admin import LocationAdmin
 from .profilecategory_admin import ProfileCategoryAdmin
 from .category_admin import CategoryAdmin
+from .coordinate_file_admin import CoordinateFileAdmin
