+++ wazimap_ng/points/admin/category_admin.py
@@ -1,153 +1,9 @@
-import json
-
-import pandas as pd
-
 from django.contrib.gis import admin
 from django.contrib.auth.models import Group
-from django import forms
-from django.utils.safestring import mark_safe
-from django.template.loader import render_to_string
-from django.conf import settings
-from django.urls import reverse
-
-from django_q.tasks import async_task
-from guardian.shortcuts import get_perms_for_model, assign_perm, remove_perm
-
-from wazimap_ng.datasets import hooks
 from wazimap_ng.general.admin.admin_base import BaseAdminModel
 
-from wazimap_ng.general.models import MetaData
-from wazimap_ng.profile.models import Profile
-from wazimap_ng.datasets.models import Licence
-
 from .. import models
-
-class InitialDataUploadChangeView(admin.StackedInline):
-    model = models.CoordinateFile
-    fk_name = "category"
-    exclude = ("task", )
-    can_delete = False
-    verbose_name = ""
-    verbose_name_plural = ""
-    extra=0
-    max_num=1
-
-
-    fieldsets = (
-        ("Uploaded Collection Data", {
-            "fields": ("document", )
-        }),
-        ("Task Details", {
-            "fields": ("get_status", "get_task_link",)
-        }),
-    )
-
-    error_and_warning_fildset = ("Error while data extraction", {
-        "fields": ("get_errors",)
-    })
-
-    readonly_fields = (
-        "document", "get_status", "get_task_link", "get_errors",
-    )
-
-    def get_status(self, obj):
-        if obj.id:
-            if obj.task:
-                return "Processed" if obj.task.success else "Failed"
-            return "In Queue"
-
-    get_status.short_description = 'Status'
-
-    def get_task_link(self, obj):
-        if obj.task:
-            task_type = "success" if obj.task.success else "failure"
-            admin_url = reverse(
-                'admin:%s_%s_change' % (
-                    obj.task._meta.app_label, task_type
-                ),  args=[obj.task.id]
-            )
-
-            return mark_safe('<a href="%s">%s</a>' % (admin_url, obj.task.id))
-        return "-"
-
-    get_task_link.short_description = 'Task Link'
-
-    def get_errors(self, obj):
-        if obj.task and not obj.task.success:
-            result = obj.task.result
-            if "CustomDataParsingExecption" in result:
-                logdir = settings.MEDIA_ROOT + "/logs/points/"
-                filename = "%s_%d_log.csv" % ("point_file", obj.id)
-                download_url = settings.MEDIA_URL + "logs/points/"
-
-                df = pd.read_csv(logdir+filename, header=None, sep=",", nrows=10, skiprows=1)
-                error_list = df.values.tolist()
-
-                result = render_to_string(
-                    'custom/render_task_errors.html', { 'errors': error_list, 'download_url': download_url + filename}
-                )
-
-            return mark_safe(result)
-        return "None"
-
-    get_errors.short_description = 'Errors'
-
-    def get_fieldsets(self, request, obj):
-        fields = super().get_fieldsets(request, obj)
-        if obj.coordinatefile_set.exists() and obj.coordinatefile_set.first().task:
-            fields = fields + (self.error_and_warning_fildset, )
-
-        return fields
-
-class InitialDataUploadAddView(admin.StackedInline):
-    model = models.CoordinateFile
-    fk_name = "category"
-    exclude = ("task", )
-    can_delete = False
-    verbose_name = ""
-    verbose_name_plural = ""
-    max_num=1
-    extra=1
-
-    fieldsets = (
-        ("Initial Data Upload - Use this form to upload file that will allow us to create points.", {
-            "fields": ("document",)
-        }),
-    )
-
-
-class CategoryAdminForm(forms.ModelForm):
-    source = forms.CharField(widget=forms.TextInput(attrs={'class': 'vTextField'}), required=False)
-    description = forms.CharField(widget=forms.Textarea(attrs={'class': 'vLargeTextField'}), required=False)
-    licence = forms.ModelChoiceField(queryset=Licence.objects.all(), required=False)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        if self.instance.id:
-            if self.instance.metadata:
-                metadata = self.instance.metadata
-                self.fields["source"].initial = metadata.source
-                self.fields["description"].initial = metadata.description
-                self.fields["licence"].initial = metadata.licence
-
-    def save(self, commit=True):
-        if self.has_changed():
-            metadata = {
-                key: self.cleaned_data.get(key) for key in [
-                    "source", "description", "licence"
-                ]
-            }
-            if not self.instance.id:
-                self.instance.metadata = MetaData.objects.create(**metadata)
-            else:
-                MetaData.objects.filter(id=self.instance.metadata.id).update(**metadata)
-        return super().save(commit=commit)
-
-    class Meta:
-        model = models.Category
-        fields = "__all__"
-
+from .forms import CategoryAdminForm
 
 @admin.register(models.Category)
 class CategoryAdmin(BaseAdminModel):
@@ -155,11 +11,10 @@ class CategoryAdmin(BaseAdminModel):
     list_filter = ("theme",)
     form = CategoryAdminForm
     exclude = ("metadata", )
-    inlines = (InitialDataUploadAddView,)
 
     fieldsets = (
         ("", {
-            'fields': ('theme', 'name', )
+            'fields': ('profile', 'theme', 'name', )
         }),
         ("Permissions", {
             'fields': ('permission_type', )
@@ -174,41 +29,6 @@ class CategoryAdmin(BaseAdminModel):
         css = {
              'all': ('/static/css/admin-custom.css',)
         }
-
-    def change_view(self, request, object_id, form_url='', extra_context=None):
-        try:
-            obj = models.Category.objects.get(id=object_id)
-            if obj and obj.coordinatefile_set.exists():
-                self.inlines = (InitialDataUploadChangeView, )
-        except models.CoordinateFile.DoesNotExist:
-            pass
-        return super().change_view(request, object_id, form_url, extra_context)
-
-
-    def save_formset(self, request, form, formset, change):
-        """
-        Given an inline formset save it to the database.
-        """
-        if formset.model == models.CoordinateFile:
-            instances = formset.save()
-            for instance in instances:
-                async_task(
-                    "wazimap_ng.points.tasks.process_uploaded_file",
-                    instance, 
-                    task_name=f"Uploading data: {instance}",
-                    hook="wazimap_ng.datasets.hooks.process_task_info",
-                    key=request.session.session_key,
-                    type="upload", assign=True, notify=True
-                )
-
-                hooks.custom_admin_notification(
-                    request.session,
-                    "info",
-                    "Data upload for %s started. We will let you know when process is done." % (
-                        instance
-                    )
-                )
-        return super().save_formset(request, form, formset, change)
 
     def save_model(self, request, obj, form, change):
         is_new = obj.pk == None and change == False
