+++ wazimap_ng/points/admin/forms/__init__.py
@@ -1,0 +1,2 @@
+from .coordinate_file_form import CoordinateFileForm
+from .category_admin_form import CategoryAdminForm
