+++ wazimap_ng/points/admin/forms/category_admin_form.py
@@ -1,0 +1,38 @@
+from django import forms
+
+from wazimap_ng.datasets.models import Licence
+from wazimap_ng.general.models import MetaData
+from ... import models
+
+
+class CategoryAdminForm(forms.ModelForm):
+    source = forms.CharField(widget=forms.TextInput(attrs={'class': 'vTextField'}), required=False)
+    description = forms.CharField(widget=forms.Textarea(attrs={'class': 'vLargeTextField'}), required=False)
+    licence = forms.ModelChoiceField(queryset=Licence.objects.all(), required=False)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        if self.instance.id:
+            if self.instance.metadata:
+                metadata = self.instance.metadata
+                self.fields["source"].initial = metadata.source
+                self.fields["description"].initial = metadata.description
+                self.fields["licence"].initial = metadata.licence
+
+    def save(self, commit=True):
+        if self.has_changed():
+            metadata = {
+                key: self.cleaned_data.get(key) for key in [
+                    "source", "description", "licence"
+                ]
+            }
+            if not self.instance.id:
+                self.instance.metadata = MetaData.objects.create(**metadata)
+            else:
+                MetaData.objects.filter(id=self.instance.metadata.id).update(**metadata)
+        return super().save(commit=commit)
+
+    class Meta:
+        model = models.Category
+        fields = "__all__"
